requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --strip-extras requirements.in
#
--index-url https://pypi.tuna.tsinghua.edu.cn/simple

aiohttp==3.9.1
    # via
    #   langchain
    #   langchain-community
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.6.0
    # via pydantic
anyio==4.2.0
    # via
    #   httpx
    #   langchain-core
    #   openai
    #   starlette
attrs==23.2.0
    # via aiohttp
certifi==2023.11.17
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via uvicorn
dataclasses-json==0.6.3
    # via
    #   langchain
    #   langchain-community
distro==1.9.0
    # via openai
dnspython==2.5.0
    # via
    #   email-validator
    #   pymongo
email-validator==2.1.0.post1
    # via pydantic
fastapi==0.109.0
    # via -r requirements.in
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.0.3
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.2
    # via httpx
httpx==0.26.0
    # via openai
idna==3.6
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
    #   yarl
jsonpatch==1.33
    # via
    #   langchain
    #   langchain-core
jsonpointer==2.4
    # via jsonpatch
langchain==0.1.4
    # via -r requirements.in
langchain-community==0.0.16
    # via
    #   -r requirements.in
    #   langchain
langchain-core==0.1.16
    # via
    #   langchain
    #   langchain-community
    #   langchain-openai
langchain-openai==0.0.5
    # via -r requirements.in
langsmith==0.0.83
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
marshmallow==3.20.2
    # via dataclasses-json
motor==3.3.2
    # via -r requirements.in
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
numpy==1.26.3
    # via
    #   langchain
    #   langchain-community
    #   langchain-openai
openai==1.10.0
    # via
    #   -r requirements.in
    #   langchain-openai
packaging==23.2
    # via
    #   langchain-core
    #   marshmallow
pydantic==2.5.3
    # via
    #   -r requirements.in
    #   fastapi
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
pydantic-core==2.14.6
    # via pydantic
pymongo==4.6.1
    # via motor
python-dotenv==1.0.1
    # via -r requirements.in
pyyaml==6.0.1
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
regex==2023.12.25
    # via tiktoken
requests==2.31.0
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
    #   langsmith
    #   tiktoken
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
    #   openai
sqlalchemy==2.0.25
    # via
    #   langchain
    #   langchain-community
starlette==0.35.1
    # via fastapi
tenacity==8.2.3
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
tiktoken==0.5.2
    # via
    #   -r requirements.in
    #   langchain-openai
tqdm==4.66.1
    # via openai
typing-extensions==4.9.0
    # via
    #   fastapi
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
urllib3==2.1.0
    # via requests
uvicorn==0.27.0
    # via -r requirements.in
yarl==1.9.4
    # via aiohttp
